requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
alembic=1.10.2=pypi_0
alpaca-ml=0.9.2=pypi_0
astroid=2.11.7=pypi_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
black=23.1.0=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7f98852_4
ca-certificates=2022.12.7=ha878542_0
certifi=2022.12.7=pyhd8ed1ab_0
charset-normalizer=3.3.2=pypi_0
click=8.1.3=pypi_0
cmaes=0.9.1=pypi_0
colorama=0.4.6=pypi_0
colorlog=6.7.0=pypi_0
coverage=7.1.0=pypi_0
cudatoolkit=11.3.1=h9edb442_10
cycler=0.11.0=pypi_0
debugpy=1.5.1=py37h295c915_0
decorator=5.1.1=pyhd8ed1ab_0
dill=0.3.7=pypi_0
dppy=0.3.2=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
exceptiongroup=1.2.0=pypi_0
ffmpeg=4.3=hf484d3e_0
flake8=5.0.4=pypi_0
fonttools=4.38.0=pypi_0
freetype=2.10.4=h0708190_1
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
greenlet=2.0.2=pypi_0
idna=3.6=pypi_0
importlib-metadata=4.2.0=pypi_0
importlib-resources=5.12.0=pypi_0
iniconfig=2.0.0=pypi_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.15.0=pyh210e3f2_0
ipython=7.33.0=py37h89c1867_0
isort=5.11.5=pypi_0
jbig=2.1=h7f98852_2003
jedi=0.18.2=pyhd8ed1ab_0
joblib=1.2.0=pypi_0
jpeg=9e=h166bdaf_1
jupyter_client=7.0.6=pyhd8ed1ab_0
jupyter_core=4.11.1=py37h06a4308_0
kiwisolver=1.4.4=pypi_0
lame=3.100=h7f98852_1001
lazy-object-proxy=1.9.0=pypi_0
lcms2=2.12=hddcbb42_0
lerc=2.2.1=h9c3ff4c_0
libdeflate=1.7=h7f98852_5
libedit=3.1.20221030=h5eee18b_0
libffi=3.2.1=1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.17=h166bdaf_0
libpng=1.6.37=h21135ba_2
libsodium=1.0.18=h36c2ea0_1
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.3.0=hf544144_1
libuv=1.43.0=h7f98852_0
libwebp-base=1.2.2=h7f98852_1
line-profiler=4.0.2=pypi_0
lz4-c=1.9.3=h9c3ff4c_1
mako=1.2.4=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.5.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mccabe=0.7.0=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py37h402132d_0
mkl_fft=1.3.1=py37h3e078e5_1
mkl_random=1.2.2=py37h219a48f_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.6=he412f7d_0
numpy=1.21.6=pypi_0
numpy-base=1.21.5=py37ha15fc14_3
olefile=0.46=pyh9f0ad1d_1
openh264=2.1.1=h780b84a_0
openjpeg=2.4.0=hb52868f_1
openssl=1.0.2u=h516909a_0
optuna=3.1.0=pypi_0
packaging=23.0=pyhd8ed1ab_0
pandas=1.1.5=pypi_0
parso=0.8.3=pyhd8ed1ab_0
pathspec=0.11.0=pypi_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=pypi_0
pip=22.3.1=py37h06a4308_0
platformdirs=3.0.0=pypi_0
pluggy=1.2.0=pypi_0
prompt-toolkit=3.0.36=pyha770c72_0
psutil=5.9.0=py37h5eee18b_0
psycopg2-binary=2.9.5=pypi_0
ptyprocess=0.7.0=pyhd3deb0d_0
py=1.11.0=pypi_0
pycodestyle=2.9.1=pypi_0
pydocstyle=6.3.0=pypi_0
pyflakes=2.5.0=pypi_0
pygments=2.14.0=pyhd8ed1ab_0
pyhdb=0.3.4=pypi_0
pylint=2.13.9=pypi_0
pyparsing=3.0.9=pypi_0
pytest=7.4.4=pypi_0
pytest-html=3.2.0=pypi_0
pytest-metadata=3.0.0=pypi_0
python=3.7.0=h6e4f718_3
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.7=2_cp37m
pytorch=1.10.0=py3.7_cuda11.3_cudnn8.2.0_0
pytorch-mutex=1.0=cuda
pytz=2022.7.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=19.0.2=py37hac76be4_2
readline=7.0=h7b6447c_5
requests=2.31.0=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.7.3=pypi_0
seaborn=0.12.2=pypi_0
setuptools=65.6.3=py37h06a4308_0
shtab=1.5.8=pypi_0
six=1.16.0=pyh6c4a22f_0
snowballstemmer=2.2.0=pypi_0
sqlalchemy=2.0.8=pypi_0
sqlite=3.33.0=h62c20be_0
termcolor=2.2.0=pypi_0
threadpoolctl=3.1.0=pypi_0
timeout-decorator=0.5.0=pypi_0
tk=8.6.11=h1ccaba5_0
tldr=3.1.0=pypi_0
tomli=2.0.1=pypi_0
torchaudio=0.10.0+cu113=pypi_0
torchbnn=1.2=pypi_0
torchvision=0.11.1+cu113=pypi_0
tornado=6.1=py37h540881e_3
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
typed-ast=1.5.4=pypi_0
typing-extensions=4.4.0=pypi_0
typing_extensions=4.5.0=pyha770c72_0
urllib3=2.0.7=pypi_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
wrapt=1.16.0=pypi_0
xz=5.2.10=h5eee18b_1
zeromq=4.3.4=h9c3ff4c_1
zipp=3.12.1=pypi_0
zlib=1.2.11=0
zstd=1.5.0=ha95c52a_0
